+++ MastersOfMana/Assets/Scripts/Spells/SpellBehaviour/Summonings/FireballBehaviour.cs
@@ -12,13 +12,13 @@ public class FireballBehaviour : A_ServerMoveableSummoning
     [SerializeField]
     private float mSpeed = 5.0f;
     [SerializeField]
-    private float mDamage = 5.0f;
+    private int mDamage = 5;
 
 	public GameObject ballMesh;
 	public GameObject explosionPrefab;
 	public float explosionRadius = 4;
 	public float explosionForce = 5;
-	public float explosionDamage = 5.0f;
+	public int explosionDamage = 5;
 
 	public TrailRenderer trail;
